navigator: relocate store interfaces from controllers/types to stores/types
ControlsStore, UIEnvironmentStore, MapPaddingStore, and Breakpoints
were declared in controllers/types.ts despite being store-shaped — no
relation to controllers. Move them to stores/types.ts next to
SessionStore/CameraStore/RouteStore/NavSheetStore. Update imports.
