+++ packages/apps/navigator/src/controllers/types.ts
@@ -14,17 +14,6 @@ export interface AppStore extends SessionStore, CameraStore, RouteStore {}
 
 export type CompassPoint = 'N' | 'S' | 'E' | 'W' | 'NE' | 'NW' | 'SE' | 'SW';
 
-export interface ControlsStore {
-  // (-180, 180] CW, 0 is north.
-  bearing: number;
-  limit: number;
-  speed: number;
-  readonly units: 'imperial' | 'metric';
-  readonly showRecenterFab: boolean;
-  readonly showRouteFab: boolean;
-  readonly showSearchFab: boolean;
-}
-
 export interface NavSheetController {
   search(query: string): Promise<SearchResult[]>;
   onSearchSelect(queryOrResult: string | SearchResult): void;
@@ -38,35 +27,4 @@ export interface NavSheetController {
   ): void;
 
   onDestinationRoutesClick(destination: SearchResult): void;
-}
-
-export interface Breakpoints {
-  readonly xs: number;
-  readonly sm: number;
-  readonly md: number;
-  readonly lg: number;
-  readonly xl: number;
-}
-
-export interface UIEnvironmentStore {
-  readonly breakpoints: Breakpoints;
-  readonly width: number;
-  readonly height: number;
-  readonly orientation: 'portrait' | 'landscape';
-  readonly isLargePortrait: boolean;
-
-  // other potential fields
-  // isTouchMobile;
-  // isDesktopPointer;
-  // isKeyboardOpen;
-}
-
-export interface MapPaddingStore {
-  readonly padding: {
-    readonly left: number;
-    readonly right: number;
-    readonly top: number;
-    readonly bottom: number;
-  };
-  readonly offset: [number, number];
 }
