+++ packages/apps/navigator/src/create-app-reactions.ts
@@ -1,9 +1,13 @@
 import type { IReactionDisposer } from 'mobx';
-import type { MapPaddingStore } from './controllers/types';
 import { wireCameraReactions } from './reactions/camera';
 import { wireRouteReactions } from './reactions/route';
 import type { MapPresenter } from './services/map-presenter';
-import type { CameraStore, NavSheetStore, RouteStore } from './stores/types';
+import type {
+  CameraStore,
+  MapPaddingStore,
+  NavSheetStore,
+  RouteStore,
+} from './stores/types';
 
 export interface ReactionDeps {
   camera: CameraStore;
