+++ packages/apps/navigator/src/stores/hooks/use-ui.ts
@@ -1,8 +1,5 @@
-import type {
-  MapPaddingStore,
-  UIEnvironmentStore,
-} from '../../controllers/types';
 import { useRootStore } from '../context';
+import type { MapPaddingStore, UIEnvironmentStore } from '../types';
 
 export function useUIEnvironmentStore(): UIEnvironmentStore {
   return useRootStore().uiEnv;
