+++ packages/apps/navigator/src/stores/types.ts
@@ -14,6 +14,48 @@ import type {
   NavPageKey,
 } from '../controllers/constants';
 
+export interface ControlsStore {
+  // (-180, 180] CW, 0 is north.
+  bearing: number;
+  limit: number;
+  speed: number;
+  readonly units: 'imperial' | 'metric';
+  readonly showRecenterFab: boolean;
+  readonly showRouteFab: boolean;
+  readonly showSearchFab: boolean;
+}
+
+export interface Breakpoints {
+  readonly xs: number;
+  readonly sm: number;
+  readonly md: number;
+  readonly lg: number;
+  readonly xl: number;
+}
+
+export interface UIEnvironmentStore {
+  readonly breakpoints: Breakpoints;
+  readonly width: number;
+  readonly height: number;
+  readonly orientation: 'portrait' | 'landscape';
+  readonly isLargePortrait: boolean;
+
+  // other potential fields
+  // isTouchMobile;
+  // isDesktopPointer;
+  // isKeyboardOpen;
+}
+
+export interface MapPaddingStore {
+  readonly padding: {
+    readonly left: number;
+    readonly right: number;
+    readonly top: number;
+    readonly bottom: number;
+  };
+  readonly offset: [number, number];
+}
+
 export interface SessionStore {
   themeMode: 'light' | 'dark';
   map: 'usa' | 'europe';
